vue2 filters 와 vue3방식의 computed 로 날짜 필터 다뤄보기
filters는 vue3에서는 사라진 문법
